requirements.txt
beautifulsoup4>=4.12.0
claude-code-sdk>=0.0.25
crawl4ai>=0.8.6
fastapi>=0.116.1
httpx>=0.28.1
langchain-anthropic>=0.3.0
langchain-core>=0.3.79
langchain-openai>=0.3.35
mcp>=1.0.0
pydantic>=2.0.0
pymilvus[milvus_lite]>=2.6.0
python-dotenv>=1.2.1
qdrant-client>=1.12.1
requests>=2.31.0
rich>=14.0.0
sentence-transformers>=5.1.0
uvicorn>=0.35.0
